(feature) Added _check script to connect /src/ files to directory
Prior to this commit, there was no way to connect the dotfiles from this
project to the home directory. The script in this commit looks for local
copies of the files in /src/ in the home directory, copies them to .backupDotfiles
if present and then creates a symbolic link to the project version of the file.